Move species being explored from products to isomers in exploreIsomer().
The act of exploring a species in a pressure-dependent reaction network
coincides with that species being moved from a unimolecular product channel
to a unimolecular isomer, as well as it being added to the list of
explored isomers. However, the former action was not being taken in
exploreIsomer(). This was causing the code to (infrequently) attempt to
explore the same isomer multiple times in a network, and thereby trip on
one of the assertions in exploreIsomer(). By explicitly moving the
species from the product channels to the isomers, we avoid the issue
entirely.

Note that, when a network is updated, its configurations are always
resorted into isomers, reactant channels, and product channels, so just
skipping a request to explore an already-explored isomer would be okay.
However, it's better to not make that request in the first place, which is
what this commit prevents.