Use smaller grain size when convoluting vibrations into density of states.
The convolution is performed using the very fast Beyer-Swinehart method.
However, in order for this to be accurate, a much smaller grain size is
necessary than for convolution using the convolve() function. In order to
keep things from being too slow, we only use the very small grain size for
convolution of the vibrational modes.

The exact value of the small grain size depends on the desired accuracy.
A choice of 10 cm^-1 indicates that we believe our vibrational frequencies
to within 10 cm^-1 or so, which should be fine for our intended application
to RMG.