Don't fail in MEASURE input if a monatomic species is given without rotors.
If you tried to specify H atom as a species and give its molecular degrees
of freedom but not its thermo (so that the former is used to compute the
latter), MEASURE would complain about the lack of rotational constants in
the input for that species. Clearly this is nonsensical for monatomic
species, so this check is now bypassed for this case only.