Fixed incorrect estimation of monatomic species thermo.
In particular, the Cp0 and CpInf of the Wilhoit fitting algorithm are
different for monatomics than for polyatomics; the former should be 2.5R
at both limits. Since monatomics are the only species with zero
frequencies, we now use that as the condition for determining monatomics
in Wilhoit.fitToDataForConstantB().

Also, in Species.generateThermoData(), we need to treat monatomics as a
special case when setting nFreq, as 3N-6 doesn't work for N = 1.