Better handling of species without indices in determining Chemkin identifier.
If no index is present, then we don't need to append "(-1)" to the
identifier, since that isn't any use. Instead we just try to use the
Species label. If it is small enough, great; otherwise raise an exception.
(Within RMG jobs there will always be an index, so this only affects
non-RMG jobs.)