Discard reaction comments from Chemkin file if we can't tell how to use them.
In principle, users can add comments to Chemkin files in any way they want,
since the Chemkin parser ignores them. However, we have been trying to
store some extra information in the comments of Chemkin files generated by
RMG. For this to work, we need to match blocks of comments with their
associated reactions. For "real world" Chemkin files, this may not always
be possible. If it is not possible, it is better to simply ignore the
comments than to fail to load the Chemkin file. This comment adds this
behavior.