Added getSumOfStates() methods to each mode in the rmgpy.statmech module.
This was done in expectation of it being more accurate to directly compute
the sum of states than to compute the density of states and integrate.
However, it turns out that any improvement in accuracy is negligible.
Nonetheless, I feel it is worth keeping, on principle at least. (At the
moment this would only affect the computation of k(E) via RRKM theory.)