Moved convolve() to new rmgpy._statmech Cython module.
This module is written in pure Cython due to the need for efficient
indexing into numpy arrays. Both the regular convolution and the Beyer-
Swinehart method have been written in the new module, which is part of
the "minimal" set of required modules to compile. This was done entirely
for speed, and it's worth it: I'd expect pdep jobs in RMG-Py to run a good
20% faster due to this change alone.