Moved Network.mapDensitiesOfStates() to fully Cythonized module _network.
Again, this was motivated by some profiling of RMG-Py jobs with pressure
dependence, which showed that this function was a significant bottleneck.
Again, the slowness is largely due to indexing of numpy arrays, which can
be sped up tremendously from pure Cython but not from Cythonized Python
code. The overall effect of this change on the speed is not as great as
before, perhaps another 10% or so speed boost. Nonetheless, it successfully
accelerates the mapDensitiesOfStates() method call so much that it no
longer appears on the profiling output at all.