Added functionality for loading settings from a file.
A deficiency of the previous input file format was the need to specify the
location on disk of the RMG database. This makes it hard to pass input
files between users (e.g. to run RMG and PopulateReactions jobs on the
website).

Instead, I've copied the matplotlib approach and created functionality for
loading settings from a file. Like matplotlib, you can store this file in
a few different places, including (1) the current working directory, (2)
a subdirectory of your home directory, and (3) the RMG-Py installation
directory. The latter is added in this commit, and should always be found,
but generally not customized, since changes will likely be overridden in
future versions.

At the moment the only setting is the path to the database on disk. You
can use environment variables in this path (since it is passed through
os.path.expandvars()). The RMG database path is now obtained from the
settings (instead of the input file) when loading RMG input files.