Print tunneling coefficients in CanTherm kinetics output.
The comments before a kinetics() block in a CanTherm output file now show
a small table with the values of k(T) with and without tunneling, plus
the values of the tunneling factors themselves, at various temperatures.
This would be useful in case anyone wanted to use the tunneling factors
in another setting, or for printing in a table in a paper, etc.