Manually add zero-point energy when not using CBS methods in CanTherm.
The energy read from this location does not include the zero-point energy;
it is the energy at the bottom of the potential well. By convention,
CanTherm always thinks in terms of ground-state energies, which by
definition include the zero-point energy.

Note that the CBS methods were using the correct ground-state energy,
and that you were supposed to be including zero-point energies in values
specified manually in the input file, so this only affects your jobs if
you were using a CBS method or manually specifying energies.