Raise exception if negative barrier encountered in Eckart tunneling calculation.
This is usually caused by users giving the wrong numbers of each atom,
which results in the E0 for the species or transition state being
referenced to the wrong reference state. Before this would usually cause
the code to get stuck in an infinite loop since the error would occur in
a Cython block; now the code will raise a ValueError and stop with a
more useful error message.