More flexibility in parsing of reactions in Chemkin files
The new implementation takes each line containing a reaction equation,
removes any comment information, extracts the three Arrhenius parameters,
and then assumes anything that is left is part of the reaction equation.
This should work for almost all of the permutations of Chemkin-like files
I have seen recently, including those with spaces in the reaction
equation and with reaction equations longer than 52 characters.

I've also added functionality for parsing reactions written as 2A=B+C
(instead of A+A=B+C) and A+B=2C (instead of A+B=2C). Since Chemkin
already requires that species identifiers begin with a letter, this is
a straightforward implementation.