New method Molecule.deleteHydrogens()
This is a bit like makeHydrogensImplicit, except that it doesn't increment the
implicitHydrogens counter on the neighboring atoms, so there is no getting them back!
This is useful for some situations where you have a shallow copy of a molecule and you want
to consider it without the hydrogens, but do not want to edit attributes of the atoms
that are shared with the original molecule.