Be careful with shallow copies of dicts of dicts! (eg. bonds matrices)
When you do a shallow copy of a dictionary of dictionaries, the inner dictionaries
have not been copied - they are the original ones. This is important when you have:
    bonds1[atom1][atom2] = bond
then
    bonds2 = bonds.copy()
is NOT a safe copy to manipulate without messing up bonds1.
It is NOT a 2-D array, but a dictionary of mutable dictionaries.
You need to copy it at least one level deep (see the diff)