qmtp module
is the module that contains the driver class of the QMTP functionality.

It also collects classes such as molFile, ThreeDMolFileCreator that were
too small to have their own module but did not really fit into one
of the other modules neither.

Many of the commments come from Greg's comment in the original RMG-Java
QMTP.java