Fix detection of paths in unittest_qm unit tests.
This way of fetching paths now works for me if I run it from the root RMG-Py folder,
eg. `python unittest_qm/inputWriterTest.py`.
Previously if I did that it wouldn't find the data directory, but if I 
ran it from inside the unittest_qm then it wouldn't find the qmtp_package modules.

Also, I'm now using the variable 'directory' instead of dir, which is the name of 
a python built-in function.