reworked qmverifier for MOPAC results
*created field with success keywords that should be in
successfully terminated output file

*created field with failure keywords, that once they are detected in
an output file, False is returned, stating that no successful MOPAC
calculation exists for this molecule.

extracted a method looking for InChIKey collisions, and false negatives
for the InChi matching between the InChI of a given molecule and the
InChI found in the MOPAC output file.