qm input writer keywords for mopac added as field
two dict's are creating storing the keywords that are added
to the QM input file based on the attempt number.

the if elif else clauses are removed. I removed Greg's comments
at each keyword attempt although they are very usefull in knowing
why he exactly chose those keywords. They should converted in unit tests
eventually.

This design should be better to be externalized completely
from the code so that the user has the power to decide which
keywords will be tried out at which attempt number.