used rmgpy's statmech module for TD properties calculation
all separate methods (except 1) of calculator are replaced by one reference
to the statesmodel. This model is composed of objects Translation,
RigidRotor and HarmonicOscillator that respectively hold info on
modes on molecular mass, moments of inertia and vibrational frequencies.

Only the chirality correction to the entropy is not contained in the
staesmodel and is therefore still present in the calculator module.