Updated references to moved rmgpy.molecule submodules.
Most of the time I think you can just import things from the subpackage
directly, e.g. ::

    from rmgpy.molecule import Molecule, Group, fromAdjacencyList

rather than trying to remember which submolecule these are defined in.
The only exception would be if you are cimporting from a Cython module;
in this case you will need to refer to the full subpackage.