Removed implicit hydrogens functionality from Molecule objects.
The intent was to conserve memory in some situations by storing the
number of hydrogen atoms adjacent to each heavy atom as an integer,
instead of storing the hydrogen atoms explicitly as Atom objects. (The
idea came from OpenBabel, which has this functionality.) However, this
turned out to be far from nontrivial to implement when dealing with
the many graph manipulation requirements of an RMG job, and I don't think
the memory savings is worth the trouble any more.

There was probably also a time savings in the graph isomorphism
evaluation when comparing implicit to implicit, but I think we can get
at least some of that savings back in other ways.