Now storing dict of edges on Vertex and vertices on Edge.
Previously we stored the edges in a dict of dicts on the Graph, and did
not store the vertices on the edge at all. However, the convenience of
having these attributes on the Vertex and Edge objects outweights the
slight increase in memory use.

This is a pretty significant change in approach, and many of the methods
of the Graph class needed modification as a result.