Added new rmgpy.molecule.adjlist module for working with adjacency lists.
This functionality is used by both Molecule and Group objects. Before it
was stored in the rmgpy.molecule.group module, a bit of an unfortunate
compromise. Now that rmgpy.molecule is a subpackage, we have space to
place the adjacency list functionality in a module of its own, as done
here.