Flux diagram creation extended focused on a centralized species
Flux diagrams can now be created that focus on a chosen species.
This is done by adding optional parameter centralSpecies to
-generateFluxDiagram
-createFluxDiagram

The parameter contains the name of the species mentioned in the chemkin
file.

If the parameter is passed to the method, then:
-the index of this particular species in speciesList is sought
-the "nodes" object is populated with the other reactants / products
in which the central species is mentioned
-creation of the edges object remains the same

The rest remains the same.