Added isIdenticalTo() and isSimilarTo() to ThermoModel class.
These methods assess the degree of similarity between two thermodynamics
models. Unlike the similar methods for the kinetics models, these do not
require the form of the models to be the same; instead, they simply
require that the models reproduce the same Cp, H, S, and G values over
a wide temperature range.