The diffModels tool now also compares the species thermo.
This seems like a logical feature, since we already compare the kinetics.
In particular, since some of the kinetic estimates depend on which
direction is exothermic, we would need to compare the thermo of the
species if the two models ever disagreed on the exothermic direction.

The compareModelSpecies() method was updated to return both matched
species in the list of common species, to better parallel the
compareModelReactions() method. The implementation of the former was
updated to match a duplicate implementation, and the latter changed into
a call to compareModelSpecies() instead.

Might need some tweaking in the future of what "identical" and "similar"
means for thermo data.