Fix ring correction searching in group additive thermo estimator.
The new method of storing bonds on atoms requires this change.
Closes #81.

NB. Ring corrections are searched using a molecule fragment with
only the ring and no adjoining atoms. If a ring correction definition
depends on ligands, it will not be found. (This is not a change in 
this commit, just an observation)