Set up to generate TS geometry estimates and send to mopac.
The geometries can now be estimated and sent to run. This
gets working code to test the TS estimates being worked on.
The current successful run test is based on the TD properties
version, so more needs to be added for the TS search. Also the 
keywords need to be expanded.