Changed use of InChI and InChiKey in QM calculations.
The full augmented InChI is now stored in 'uniqueIDlong' parameter
on a QMMolecule's Geometry. This is needed for tests later on to ensure
the right molecule was found. The uniqueID is still the augmented InChI KEY.

The MOPAC input file now contains this augmented InChI as the title.

Renamed getinchiaug to getinchikeyaug (which is what it does)