Fix missing (and wrong) units in QM-generated ThermoData.
Beware. RMG-Py tends to think in SI, and that means J/mol for H, not kJ/mol. Certainly not kcal/mol.
Best to be explicit, and careful. I wonder if this could be enforced (eg. the Quantity data type could
check for missing units?) without too much of a performance hit.