Use correct units for ThirdBody et al. when loading old database.
This updates the function for loading the old (RMG-Java) database so that
it correctly parses ThirdBody, Lindemann, and Troe reactions, including
using arrheniusLow instead of arrheniusHigh for ThirdBody reactions -
now detected using the +M species - and using one reaction order higher
to assess the units of the low-pressure limit expressions. I was able to
re-import enough of the old database to show that this seems to work as
intended.

Closes #96.