Added a hacky script to process the Thermodynamics comparison test.
Having run the RMG-Java ThermoDataEstimator on the input file generated by 
the script in the previous commit, this script should compare the output
with the .thermo files.

Also tweaked the makeThermoTester script to allow mopac calculations on radicals.
Because RMG-Py isn't doing HBI method for radicals yet, this should allow 
a more direct comparison.