Changed PDep checking in KineticsLibrary.saveOld
With the new criteria for importing duplicates from Java (f21a1ef42e75578193cc924d3f727747564f20f2),
MultiKinetics objects may contain kinetics from both PDep and non-
PDep reactions as long as they share the same reactants, products,
and reversibility.

This may not be the desired behavior in the long run, but if it is,
this fix allows the individual kinetics entries to be exported
back to their appropriate locations without raising errors.