Moved physical constants to new rmgpy.constants module.
This gives users the option of importing the constants separately from
the Quantity functionality. The constants have also been made into
module-level variables instead of attributes of an otherwise-needless
class. The same globals() trick allows the defined module-level constants
to be accessible from both Python and Cython contexts; when cimported,
the C version of the constants will be made available to Cython modules.