Updated import statements to find new rmgpy.constants module.
Many of the imports of the rmgpy.quantity module only asked for one or
the other of the Quantity class or the constants object, so there seems
to be at least a small amount of sense in decoupling them. Still, many
other imports wanted both anyway, so this may just be a cosmetic change.