Improved classes for representing physical quantities.
There are now separate classes for representing scalar and higher-order
quantities, aptly named ScalarQuantity and ArrayQuantity, respectively.
For the most part, you don't need to use these directly, though; instead,
use the Quantity() function, which returns either a ScalarQuantity or
ArrayQuantity object based on the input value, units, and uncertainty.

The value attribute of both ScalarQuantity and ArrayQuantity has been
changed to a property that returns the value of the quantity in the
given units. Previously it had stored the value in the SI equivalent to
facilitate speedy computations, which worked but was rather awkward to
explain to people. What was the value attribute is now called value_si
to emphasize that it stores the value in the SI equivalent. This is
still available directly for speedy computations.

I've also revamped the unit tests accordingly.