Completed Gaussian writing, runing, and parsing output.
The units of the outputs need to be checked to ensure the CCLib
parser returns numbers in the same units as the CCLib mopac
parser. It looked fine after a quick overview of the numbers.
Also, a check and nice error message needs to be added if the
gaussian executable path is set incorrectly.