Added handling for absent KineticsFamily.reverse attribute
The reverse attribute, which stores the name of a reaction family's
opposite, was added in f566b2526a7523504a609a45367bc917d3b7fb9d, but the RMG-database branch that
actually introduces this attribute to each family is not yet ready
to be pulled in. Until then, this except block allows families to
load without error.