Moved rmgpy.constants unit tests to rmgpy package.
The motivation here is to have the unit tests live in modules alongside
the modules they test. This makes them more visible when editing the
codebase, which encourages developers to keep them up to date. We can
also use nose to automatically search and run the unit tests, perhaps as
a post-commit hook to further prevent the tests from becoming stale.