Fix to allow blank rules depositories
In light of my effort to migrate rate rules back to their original
NIST entries (and, eventually, into training rates), it is now
possible that a family can be entirely devoid of rate rules. This
tweak prevents crashing upon trying to determine the appropriate
index for training entries to be added back to these blank rule
depositories.