Added functions for validation of various unit types.
One thing that's been missing from our handling of physical quantities
has been the enforcing of unit types. For example, there is nothing
currently stopping someone from storing a moment of inertia in units of
cm^-1, which would lead to crazy rotational and torsional partition
functions, etc. This commit provides the UnitType class, a functor
(a function with associated metadata) for creating ScalarQuantity or
ArrayQuantity objects of a unit type. Several module-level instances of
UnitType are also defined, e.g. Inertia for moments of inertia.

The UnitType objects also give us the opportunity to handle unusual
units, e.g. frequencies in units of cm^-1 or K instead of s^-1, etc.
Note that RateCoefficient is currently implemented as a special case, as
these can have different units depending on the reaction order.

Also note that all energies, enthalpies, free energies, entropies, heat
capacities, concentrations, fluxes, and rate coefficients are treated as
intensive. This means that you must use mol or molecule for these. All
other unit types are extensive. A special case is added for mass to
convert e.g. g/mol to amu.

There are a large number of unit tests to ensure that the unit types
are acting as intended, as there are several edge cases.