Use MathJax instead of PNG files in generated Sphinx documentation.
MathJax has at least three advantages: (1) no LaTeX installation
required to compile the HTML documentation, (2) documentation builds
much faster, and (3) generated equations scale nicely in the browser.
The only catch is that you need an internet connection to be able to see
the equations in the generated HTML, which shouldn't be a problem.