Added Mode class to new rmgpy.statmech.mode module.
The Mode class is a base class for all representations of molecular
degrees of freedom. It provides a place to collect common attributes and
methods. For example, the quantum attribute is used to indicate whether
we wish to use the classical or quantum mechanical model for this degree
of freedom.