Added LinearRotor class to new rmgpy.statmech.rotation module.
The LinearRotor class represents a rigid rotor for a linear molecule, i.e.
a rotor with a single moment of inertia (or rotational constant). Methods
are available for computing thermodynamic and statistical mechanical
properties of the rotor using either classical or quantum approaches.
The classical results are much faster to evaluate and nearly always
accurate except at temperatures well below room temperature.

The LinearRotor class inherits from a new Rotation base class, which
itself inherits from the Mode class and adds an attribute for the
symmetry number of the rotational degree of freedom.