Added NonlinearRotor class to rmgpy.statmech.rotation module.
Nonlinear rotors have multiple (usually three) moments of inertia that
are used to compute the thermodynamic and statistical mechanical
contributions from the rotor. As with the linear rotors, the classical
nonlinear rotor is almost always a reasonable choice (except at very low
temperatures). For nonlinear rotors the quantum mechanics solution is not
straightforward, and has not been implemented at this time; you will get
a NotImplementedError if you try to use it.