Added HinderedRotor class to rmgpy.statmech.torsion module.
The HinderedRotor class represents a one-dimensional hindered rotor,
described by a reduced moment of inertia and an angle-dependent potential
function. There are two options for representing the potential: a simple
cosine expression and a more general Fourier series.

Either the classical or quantum mechanical solutions can be used for either
potential, except that the sum and density of states cannot be expressed
classically for the Fourier series potential. A semiclassical correction
to the partition function is also available, although currently not used
by default.