Added Conformer class to rmgpy.statmech.conformer module.
The Conformer class represents an individual conformation on a potential
energy surface. Each configuration is defined by a set of molecular
degrees of freedom (translation, rotation, vibration, and/or torsion),
an electronic spin multiplicity, and a number of optical isomers. The
class provides convenience methods for evaluating the thermodynamic and
statistical mechanical properties of the configuration from these
parameters.

In general, each conformer corresponds to a unique geometry (positions
of each atom). Storing that geometry makes it possible to extract some
statistical mechanical parameters (e.g. total mass and moments of
inertia) directly. Several methods for doing so have been provided.

Note that, also the ground-state energy E0 is stored as an attribute on
the Conformer object, it is not currently used to evaluate the
thermodynamic of statistical mechanical properties. This is because we
prefer to work with relative rather than absolute energies so as to not
care about the chosen zero of energy.