Added chapter to developers guide for rmgpy.statmech subpackage.
This chapter describes the classes and functions defined in the
rmgpy.statmech subpackage. Mostly this is extracting the class and
function docstrings using the sphinx.ext.autodoc functionality. Note that
we have used the embedsignature=True option in the Cython modules to
improve the look of the generated Sphinx documentation.

Generally equations in docstrings require either escaped backslashes or
raw docstrings. Neither of these is ideal, so it's probably better to
put most equations in the .rst files if possible.