Added HeatCapacityModel class to rmgpy.thermo.model module.
The HeatCapacityModel class represents a thermodynamic model of heat
capacity (and by extension enthalpy, entropy, and free energy). It
provides a place to store attributes and methods common to all heat
capacity models, including the minimum and maximum valid temperature of
the model and a string comment. Methods for getting the heat capacity,
enthalpy, entropy, and Gibbs free energy are enumerated; these must be
overridden in classes derived from HeatCapacityModel.