Added Wilhoit class to rmgpy.thermo.wilhoit module.
The Wilhoit class provides a heat capacity model based on the Wilhoit
polynomial. This expression has the advantages of giving the correct heat
capacity values in the limits of zero and infinite temperature, as well
as providing a very reasonable representation for intermediate
temperatures. The tradeoff is that the expression is more expensive to
evaluate than other heat capacity models.