Added NASA class to rmgpy.thermo.nasa module.
The NASA class provides a heat capacity model based on a simple polynomial
in temperature. This expression is very efficient to evaluate, but has no
real physical basis, and therefore may not always be the most accurate.
To compensate, often a pair of polynomials is used. The NASA class allows
for use of one, two, or three polynomials -- each of seven or nine NASA
coefficients -- to represent the heat capacity, enthalpy, entropy, and
free energy. Each individual polynomial is represented using the
NASAPolynomial helper class.