Added methods for converting between heat capacity models.
The methods are toThermoData(), toWilhoit(), and toNASA() as appropriate.
When converting to NASA, you must specify the desired valid temperature
range and a (fixed or initial guess for) the Tint value. When converting
from NASA to Wilhoit or ThermoData, you must specify Cp0 or CpInf
explicitly. A number of helper methods for computing integrals needed by
the Wilhoit-to-NASA conversion functionality have also been added to the
Wilhoit and NASAPolynomial objects.