Added KineticsModel and PDepKineticsModel to rmgpy.kinetics.model module.
The KineticsModel class represents a generic kinetics model that is a
function of temperature only. It provides a place to store attributes and
methods common to all kinetics models, including the minimum and maximum
valid temperature of the model, the reaction order, and a string comment.
A method for getting the rate coefficient k(T) is enumerated; this must
be overridden in classes derived from KineticsModel.

The PDepKineticsModel class represents a generic kinetics model that is a
function of both temperature and pressure. In addition to the above, it
provides attributes for storing the minimum and maximum valid pressure of
the model, and a set of collision efficiencies used to modify the
effective pressure for various bath gases.