Added PDepArrhenius class to rmgpy.kinetics.arrhenius module.
The PDepArrhenius class represents a model of pressure-dependent kinetics
based on a set of Arrhenius fits at several pressures. The k(T,P) value
is evaluated by logarithmic interpolation between the nearest pair of
Arrhenius expressions to the desired pressure. For simple
pressure-dependent systems this can be okay, but often the complexity of
the system makes the observed kinetics very non-Arrhenius in behavior.