Added Chebyshev class to rmgpy.kinetics.chebyshev module.
The Chebyshev class represents a model of pressure-dependent kinetics
based on a set of Chebyshev polynomials. This model has the advantage of
being able to fit a wide range of k(T,P) behavior; however, there is no
meaningful physical interpretation of the polynomial-based fit.

There is currently no k(T,P) model that consistently gives both a good fit
and a meaningful physical interpretation for complex pressure-dependent
systems. Of the two criteria, I believe the first to be more important;
that's why I generally recommend Chebyshev to anyone who asks.