Added ThirdBody, Lindemann, Troe classes to rmgpy.kinetics.falloff module.
The ThirdBody class represents kinetics in the low-pressure limit, where
a third body is always required to obtain the necessary energy for
reaction.

The Lindemann class represents the Lindemann model of falloff, which was
the first to qualitatively model the effect of pressure-dependent
falloff. However, the Lindemann model is generally not quantitative due
to its oversimplicity.

The Troe formulation attempts to make the Lindemann model quantitative
through introduction of a temperature-dependent broadening factor F.
Several additional parameters are required to compute F. For simple
falloff reactions the Troe formulation is often adequate; however, in
complex, multi-well reaction networks the Troe formulation is
inappropriate, especially for well-skipping reactions.