Added Wigner class to rmgpy.kinetics.tunneling module.
The Wigner class represents a model of tunneling based on the Wigner
formulation. This model represents the first correction term in a
perturbative expansion for a parabolic barrier, and is therefore only
accurate in the limit of a small tunneling correction. Furthermore, the
Wigner model does not have a well-defined microcanonical tunneling
function. For these reasons, the Wigner model is provided largely for
historical reasons.