Added Eckart class to rmgpy.kinetics.tunneling module.
The Eckart class represents a model of tunneling based on the Eckart
potential, which has fitted parameters for the reactant energy, product
energy, and the imaginary frequency of motion along the reaction
coordinate. This simple potential admits an analytical expression for
the microcanonical tunneling function, which can be numerically
integrated to obtain the canonical tunneling factor. The Eckart model
is often a reasonable estimate of the tunneling contribution, and can
give very large correction factors at low temperature.