Updated versions of Species and TransitionState classes.
These classes now use the Conformer object in place of the old
StatesModel object. Since E0 is stored on the Conformer object, it is
no longer stored on the Species or TransitionState object. For Species,
HeatCapacityModel replaces ThermoModel, and molecularWeight becomes a
scalar quantity wrapped in a property.

Both Species and TransitionState also add methods for computing the
thermodynamic and statistical mechanical properties. For the thermo
properties, the methods use the HeatCapacityModel if present, then the
conformer if present, then raise an exception if neither are found.