Updated version of Reaction class.
The new Reaction class adds a label attribute and removes the thirdBody
attribute; the latter is not needed thanks to the ThirdBody class for
representing kinetics. The methods for computing k(T) using TST and
k(E) using ILT or RRKM theory have been updated; the latter have been
moved to a new (not yet committed) location. The methods for computing
the enthalpy, entropy, and free energy of reaction now call the
corresponding methods of the Species object, rather than using each
species' thermo attribute directly.