Added SingleExponentialDown class to rmgpy.pdep.collision module.
The SingleExponentialDown class represents a single exponential down
model of collisional energy transfer. This is the simplest such model,
as it only requires one parameter: the average energy transferred in a
deactivating collision. (This can also be given as a power law in
temperature, which requires three parameters.) A method for generating
the matrix of collisional energy transfer probabilities is provided.