Added getDensityOfStatesForst() function to rmgpy.statmech.conformer module.
This function provides the ability to compute the density of states
using the method of steepest descents to apply the inverse Laplace
transform to the partition function. This approach has the advantage of
being able to handle any degree of freedom that has an analytical
partition function; an analytical density of states is not required.
(In particular, we can use semiclassical hindered rotors.) It is also
generally not any less accurate than direct count methods, except that
it applies continuous smoothing that is an approximation (but a very
good one) at low energies. We want continuous densities of states for
k(T,P) calculations anyway.

Despite the name, the function actually computes the sum of states first,
as it is more numerically stable, and then the density of states. The
first-order steepest descents approximation is used by default instead
of the second-order approximation due to its better numerical stability;
the accuracy tradeoff is less than 5%.