Added Conformer.getActiveModes() function.
This function returns a list of the active modes of a molecular
conformer. This list always includes all internal degrees of freedom
(vibrational and torsional), and includes all or part of the external
rotational modes as determined by the activeJRotor and activeKRotor
parameters. If either or both of the J-rotor or K-rotor are not active,
then the NonlinearRotor is approximated as a symmetric top (using a
LinearRotor and a KRotor.)

If you are solving the 1D master equation to get k(T,P), you want both
the J-rotor and K-rotor to be active. This is less accurate at lower
temperatures. If you want the J-rotor to be adiabatic (as is physically
more reasonable at low temperatures) you must solve the 2D master
equation, which is expensive.