Added Configuration class to rmgpy.pdep.configuration module.
The Configuration class represents a molecular configuration on a
potential energy surface. It can be used for a unimolecular isomer,
bimolecular reactant or product channel, or a transition state. Methods
are provided for getting the thermodynamic and statistical mechanical
properties. Methods for mapping the density and sum of states from
energies referenced to the zero-point energy to energies references to
some other energy are provided; interpolation is used where needed to
do this.