Added rmgpy.pdep.msc module for the modified strong collision method.
This is a fast, robust, and often reasonably accurate method of getting
the k(T,P) values for a pressure-dependent reaction network. It has a
tendency to underpredict the effect of falloff for larger reaction
networks, but in general works well over a wide range of conditions
due to the use of a collision efficiency (the "modified" in modified
strong collision).