Added rmgpy.pdep.rs module for the reservoir state method.
Only a bit more expensive than the MSC method, the RS method is often
more accurate at lower temperatures. However, the approximations used in
the RS method fail terribly at high temperatures, which makes it far
less robust than the MSC method. (It may be possible to correct for
this, but the current implementation does not.)